Refactored LibraryEntry -> ILibraryEntry; ILibraryMetaEntry
The processing code was moved to static factory-like loaders; and the primary data model was reduced and made into an interface. The meta-entry is meant to facilitate the loading of relevant data from multiple data-sources without any ties to the file system or a data-source ID (such as a file name or database record ID).

MusicBrainz nuget api installed and ready for consumption.