Adding open / save. couple issues.
Bringing in serializer from another project. If anyone knows a good solution for a binary serializer that works with Avalonia, please contact me. I have a working "RecursiveSerializer" that replaces .NET and does not require attribute tags. This will work in a variety of situations - including where you need to attach to a heavy object (such as Avalonia's ObservableObject). So, I'm going to start a couple new repositories to try and complete the packaging for mine.. but it's a lot of extra work.